requirements.txt
langgraph
langchain
langchain-openai
supabase
python-docx
gspread
google-auth
requests
python-dotenv
openpyxl
reportlab
